Feature/add branch movement

+++ database/knex-create-tables.js
@@ -1,10 +1,12 @@
 const knex = require('knex')(require('../knexfile').development);
 const bcrypt = require("bcrypt");
 const knexCore = require('./knex-core');
+require('dotenv').config();
+const path = require("path");
+const fs = require('fs');
 
 const dropTable = table => { 
-    console.log (`dropping ${table}`);
-    return knex.schema.dropTable(table).catch(err => console.log (`could not drop ${table}`, err))};
+    return knex.schema.dropTable(table).catch(err => console.error (`could not drop ${table}`, err))};
 
 const dropAllTables = () => {
 
@@ -13,12 +15,11 @@ const dropAllTables = () => {
     .then (dropTable('trees'))
     .then (dropTable('users'))
     .catch(err => {
-        console.log ("error dropping tables", err);
+        console.error ("error dropping tables", err);
     })
 }
 
 const createUsersTable = () => {
-    console.log ("Creating users table...");
     return knex.schema
     .createTable('users', (table) => {
         table.bigincrements('user_id').unsigned().notNullable()
@@ -32,16 +33,14 @@ const createUsersTable = () => {
     .catch(err => {
         switch (err.errno) {
             case 1050:
-                console.log ("users table already exists");
                 break;
             default:
-                console.log ('create user table error', err);
+                console.error ('create user table error', err);
         }
     })
 }
 
 const createTreesTable = () => {
-    console.log (`creating tree table`)
     return knex.schema
     .createTable('trees', (table) => {
         table.bigincrements('tree_id').unsigned().notNullable()
@@ -56,17 +55,14 @@ const createTreesTable = () => {
     .catch(err => {
         switch (err.errno) {
             case 1050:
-                console.log ("trees table already exists");
                 break;
             default:
-                console.log ('create trees table error', err.errno);
+                console.error ('create trees table error', err.errno);
         }
     })
 }
 
 const createBranchesTable = () => {
-    console.log ("Creating branches table\n\n\n\n");
-
     return knex.schema
     .createTable('branches', table => {
         table.bigincrements('branch_id').unsigned().notNullable()
@@ -77,10 +73,9 @@ const createBranchesTable = () => {
     .catch(err => {
         switch (err.errno) {
             case 1050:
-                console.log ("branches table already exists");
                 break;
             default:
-                console.log ('create branches table error', err.errno);
+                console.error ('create branches table error', err.errno);
         }
     });
 }
@@ -88,17 +83,17 @@ const createBranchesTable = () => {
 const createModulesTable = () => {
     return knex.schema
     .createTable('modules', table => {
-        table.string('module_name').primary().notNullable()
+        table.bigincrements('id').primary()
+        table.string('module_name').notNullable()
         table.string('module_icon', 2048).notNullable()
         table.unique('module_name')
     })
     .catch(err => {
         switch (err.errno) {
             case 1050:
-                console.log ("modules table already exists");
                 break;
             default:
-                console.log ('create modules table error', err.errno);
+                console.error ('create modules table error', err.errno);
         }
     });
 }
@@ -112,10 +107,9 @@ const createModuleTable = moduleName => {
     .catch(err => {
         switch (err.errno) {
             case 1050:
-                console.log (`${moduleName} table already exists`);
                 break;
             default:
-                console.log (`create ${moduleName} table error1`, err.errno);
+                console.error (`create ${moduleName} table error1`, err.errno);
         }
     });
 }
@@ -130,7 +124,6 @@ const insertUser = (userName, password, email) => {
 }
 
 const addModule = (moduleName, moduleIcon) => {
-    console.log(`knex-create-tables.js addModule (${moduleName}, ${moduleIcon})`);
     return knex('modules')
     .insert({
         module_name: moduleName,
@@ -143,8 +136,6 @@ const addModule = (moduleName, moduleIcon) => {
 }
 
 const addUser = (userName, password, email) => {
-
-    console.log (`knex-create-tables.js addUser ${userName}`);
     let userId;
     let reservedTree = 1;
     let branchPool = [];
@@ -152,7 +143,6 @@ const addUser = (userName, password, email) => {
     insertUser (userName, password, email)
     .then(info => {
         userId = info[0];
-        console.log(`User id from ${userName} is ${userId}`)
         return knexCore.createNewBranch(reservedTree)
     })
     .then(info => {
@@ -165,7 +155,6 @@ const addUser = (userName, password, email) => {
     })
     .then(info => {
         branchPool.push(info[0]);
-        console.log('branchPool', branchPool)
         return info[0]
     })
     .then(info => {
@@ -178,16 +167,14 @@ const addUser = (userName, password, email) => {
         })
     })
     .then(info => {
-        console.log(`User ${userName} has been added.`);
     })
     .catch (err => {
         switch (err.errno) {
             case 1062:
-                console.log (`${userName} already exists in users table`);
                 break;
             
             default:
-                console.log ('addUser error', err);
+                console.error ('addUser error', err);
         }
     })
 
@@ -200,18 +187,33 @@ const createReservedBranch = treeId => {
     })
 }
 
-exports.createTables = () => {
+exports.createTables = (req, res) => {
+    if (!req.query.key) {
+        return res.status(401).json({message: 'missing key'});
+    }
+
+    const {key} = req.query;
+
+    if (key !== process.env.INITIALIZEKEY) {
+        return res.status(401).json({message: 'invalid key'});
+    }
+
+    // directory path
+    const dir = path.resolve('assets/');
+
+    // delete directory recursively
+    fs.rm(dir, { recursive: true }, (err) => { });
 
     let reservedTree = '';
     let branchPool = [];
 
-    dropTable('quill')
+    dropTable('notes')
     .then(info => {return dropTable('videos')})
     .then(info => {return dropTable('chat')})
     .then(info => {return dropTable('post')})
     .then(info => {return dropTable('documents')})
     .then(info => {return dropTable('webpage')})
-    .then(info => {return dropTable('image_gallery')})
+    .then(info => {return dropTable('images')})
     .then(info => {return dropTable('modules')})
     .then(info => {return dropTable('branches')})
     .then(info => {return dropTable('trees')})
@@ -220,21 +222,23 @@ exports.createTables = () => {
     .then(info => {return createTreesTable()})
     .then(info => {return createBranchesTable()})
     .then(info => {return createModulesTable()})
-    .then(info => {return knex('users').insert({user_name: 'system', password: 'asdghaskghalhewieufdsvagksajegf', email: 'noemail@noemail.com'})}) // IMPORTANT: Move to .env
+    .then(info => {return knex('users').insert({user_name: 'system', password: 'asdghaskghalhewieufdsvagksajegf', email: 'noemail@noemail.com'})}) // Note: Hackers cannot use this info to bypass because a password hash will be applied when logging in. In other words, this value is fake hash, not a fake password. Thus, it is not in the .env file.
     .then(info => {
         // create reserved tree to be assigned to branches in the users' branch pool
-        console.log ('reserved tree', info);
         return knexCore.initializeNewTree(info[0], '/svg/tree.svg', 'reserved system tree', 'This tree is used as a reference for branches that are assigned to each user branch pool', '#000000', 1)
     })
-    .then(info => {return addModule('quill', '/svg/quill.svg')})
-    .then(info => {return addModule('image_gallery', '/svg/image_gallery.svg')})
+    .then(info => {return addModule('notes', '/svg/quill.svg')})
+    .then(info => {return addModule('images', '/svg/image_gallery.svg')})
+    .then(info => {return addModule('videos', '/svg/video.svg')})
     .then(info => {return addModule('documents', '/svg/documents.svg')})
     .then(info => {return addModule('chat', '/svg/chat.svg')})
     .then(info => {return addModule('post', '/svg/post.svg')})
     .then(info => {return addModule('webpage', '/svg/webpage.svg')})
-    .then(info => {return addModule('videos', '/svg/video.svg')})
-    .then(info => {return addUser('admin', "Technologist@33301", 'michaelwood33311@icloud.com')}) // IMPORTANT: Move password to .gitignored .env
+    .then(info => {return addUser('admin', process.env.ADMINPASSWORD, process.env.ADMINEMAIL)})
+    .then(info => {
+        res.status(200).json({message: 'tables created'});
+    })
     .catch (err => {
-        console.log ("Create Tables Error:", err);
+        res.status(500).json({error: err});
     })
 }
